# openapi.yaml
openapi: 3.0.0
info:
  title: reciclaapp API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3016/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /transaccion:
    get:
      tags:
        - Transaccion
      summary: Listar todos los Transaccions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Transaccions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Transaccion
      summary: Crear Transaccion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Transaccion creado
  /transaccion/{id}:
    get:
      tags:
        - Transaccion
      summary: Obtener Transaccion por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaccion encontrado
        '404':
          description: Transaccion no encontrado
    patch:
      tags:
        - Transaccion
      summary: Actualizar Transaccion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Transaccion actualizado
    delete:
      tags:
        - Transaccion
      summary: Eliminar Transaccion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaccion eliminado
  /transaccion/{id}/upload-image:
    post:
      tags:
        - Transaccion
      summary: Subir imagen a Transaccion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
  /recompensa:
    get:
      tags:
        - Recompensa
      summary: Listar todos los Recompensas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Recompensas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Recompensa
      summary: Crear Recompensa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Recompensa creado
  /recompensa/{id}:
    get:
      tags:
        - Recompensa
      summary: Obtener Recompensa por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recompensa encontrado
        '404':
          description: Recompensa no encontrado
    patch:
      tags:
        - Recompensa
      summary: Actualizar Recompensa
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Recompensa actualizado
    delete:
      tags:
        - Recompensa
      summary: Eliminar Recompensa
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recompensa eliminado
  /recompensa/{id}/upload-image:
    post:
      tags:
        - Recompensa
      summary: Subir imagen a Recompensa
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida